added mapping for the functions sum, max, min, join
