+++ src/utils/form/extractKoboFields.ts
@@ -1,3 +1,6 @@
+import { mapKoboExpression } from './kobo/mapKoboExpression';
+import { commonProperties } from './kobo/commonProperties';
+
 /**
  * Saves all groupElement questions equivalent to the panel question.
  * Groups and panels can be nested.
@@ -54,158 +57,6 @@ const AVAILABLE_TYPES = [
   'begin_repeat',
   'end_repeat',
 ];
-
-/**
- * Maps expressions from kobo questions to a expression format that will work on the SurveyJS.
- *
- * The numeric operators (Greater than >, Less than <, Greater than or equal to >=, Less than or equal to <=) are used the in same way in Kobo and SurveyJS.
- *
- * @param koboExpression the initial kobo logic expression
- * @param questionName name of the question to replace in the expression
- * @returns the mapped logic expression that will work on the SurveyJS form
- */
-const mapKoboExpression = (koboExpression: string, questionName?: string) => {
-  // Replace . with {questionName}
-  if (questionName) {
-    // Expressions in Kobo can have ' . ' to indicate that the expression is about the question in which it is defined.
-    // Example: a Validation Criteria can be ". > 5 ": the value of the question itself must be greater than 5
-    koboExpression = koboExpression.replace(/\./g, `{${questionName}}`);
-  }
-  // Not contains
-  koboExpression = koboExpression.replace(
-    /not\(selected\(\$\{(\w+)\}, '(.*?)'\)\)/g,
-    "{$1} notcontains '$2'"
-  );
-  // Replace not(...) with !(...)
-  koboExpression = koboExpression.replace(/not\(/g, '!(');
-  // Replace mod with %
-  koboExpression = koboExpression.replace(
-    /([^\/]*)\s+mod\s+([^\/]*)/g,
-    (match, before, after) => {
-      const transformedBefore = before.replace(/\$\{(\w+)\}/g, '{$1}');
-      const transformedAfter = after.replace(/\$\{(\w+)\}/g, '{$1}');
-      return `${transformedBefore} % ${transformedAfter}`;
-    }
-  );
-  // Replace div with /
-  koboExpression = koboExpression.replace(
-    /([^\/]*)\s+div\s+([^\/]*)/g,
-    (match, before, after) => {
-      const transformedBefore = before.replace(/\$\{(\w+)\}/g, '{$1}');
-      const transformedAfter = after.replace(/\$\{(\w+)\}/g, '{$1}');
-      return `${transformedBefore} / ${transformedAfter}`;
-    }
-  );
-  // Empty
-  koboExpression = koboExpression.replace(/\$\{(\w+)\} = ''/g, '{$1} empty');
-  // Equal to
-  koboExpression = koboExpression.replace(
-    /\$\{(\w+)\} = '(.*?)'/g,
-    "({$1} = '$2')"
-  );
-  // Contains
-  koboExpression = koboExpression.replace(
-    /selected\(\$\{(\w+)\}, '(.*?)'\)/g,
-    "{$1} contains '$2'"
-  );
-  // No empty
-  koboExpression = koboExpression.replace(
-    /\$\{(\w+)\} != ''/g,
-    '{$1} notempty'
-  );
-  // Not equal to
-  koboExpression = koboExpression.replace(
-    /\$\{(\w+)\} != '(.*?)'/g,
-    "{$1} <> '$2'"
-  );
-  // Replace ends-with with endsWith
-  koboExpression = koboExpression.replace(/ends-with\(/g, 'endsWith(');
-  // Replace indexed-repeat(...) with indexedRepeat(...) and handle the first parameter to be the question/field name and not the question value
-  koboExpression = koboExpression.replace(
-    /indexed-repeat\(\$\{(\w+)\},\s*\$\{(\w+)\},\s*(.*?)\)/g,
-    'indexedRepeat($1, {$2}, $3)'
-  );
-  // Replace position(..) with {panelIndex}
-  koboExpression = koboExpression.replace(/position\(\.\.\)/g, '{panelIndex}');
-  // Replace count-selected with length
-  koboExpression = koboExpression.replace(/count-selected\(/g, 'length(');
-  // Replace of format-date-time with formatDateTime
-  koboExpression = koboExpression.replace(
-    /format-date-time\(/g,
-    'formatDateTime('
-  );
-  // Replace if with iif
-  koboExpression = koboExpression.replace(/if\(/g, 'iif(');
-  // TODO: FIX not working with expressions like if(${number1} + ${number2} > 10,45,30) + today()
-  // For calculations with today() + or - days, add addDays() custom function to work on oort
-  koboExpression = koboExpression.replace(
-    /today\(\)\s*([\+\-])\s*(\w+)/g,
-    (match, operator, term) => {
-      const transformedTerm = term.replace(/\$\{(\w+)\}/g, '{$1}');
-      if (operator === '+') {
-        return `addDays(today(), ${transformedTerm})`;
-      } else {
-        return `addDays(today(), -${transformedTerm})`;
-      }
-    }
-  );
-  // Replace now() with currentDate()
-  koboExpression = koboExpression.replace(/now\(\)/g, 'currentDate()');
-  // Date values
-  koboExpression = koboExpression.replace(/date\('(.*?)'\)/g, "'$1'");
-  // Replace any remaining ${variable} to {variable}
-  koboExpression = koboExpression.replace(/\$\{(\w+)\}/g, '{$1}');
-  return koboExpression;
-};
-
-/**
- * Extract from a Kobo question the constraint properties and return the validators property for an SurveyJS question
- *
- * @param question Kobo question object
- * @returns validators property for an SurveyJS question
- */
-const validators = (question: any) => {
-  return {
-    validators: [
-      {
-        type: 'expression',
-        text: question.constraint_message
-          ? typeof question.constraint_message === 'string'
-            ? question.constraint_message
-            : question.constraint_message[0]
-          : '',
-        expression: mapKoboExpression(question.constraint, question.$autoname),
-      },
-    ],
-    validateOnValueChange: true,
-  };
-};
-
-/**
- * Extract from a Kobo question the common properties in a object for a SurveyJS question
- *
- * @param question Kobo question object
- * @param type type of the questions
- * @param title optional title
- * @returns the common properties in a object for a SurveyJS question extracted from the Kobo question
- */
-const commonProperties = (question: any, type: string, title?: string) => {
-  return {
-    type,
-    name: question.$autoname,
-    title: title ?? question.label[0],
-    valueName: question.$autoname,
-    isRequired: question.required,
-    ...(question.hint && { description: question.hint[0] }),
-    ...(question.default && {
-      defaultValue: mapKoboExpression(question.default),
-    }),
-    ...(question.relevant && {
-      visibleIf: mapKoboExpression(question.relevant),
-    }),
-    ...(question.constraint && validators(question)),
-  };
-};
 
 /**
  * Get all the groups names that are related to a kobo element.
